refactor(tools): move function validate_labels from tools.utils to conf_loader
KAG-3094
